Moved simplified export manager to IO module
